requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
